completing resource based authorization from the course

+++ FrontLineCleaners.Application/Commands/DeleteCleanerCommandHandler.cs
@@ -1,13 +1,15 @@
 ﻿using AutoMapper;
 using FrontLineCleaners.Domain.Entities;
 using FrontLineCleaners.Domain.Exceptions;
+using FrontLineCleaners.Domain.Interfaces;
 using FrontLineCleaners.Domain.Repositories;
 using MediatR;
 using Microsoft.Extensions.Logging;
 
 namespace FrontLineCleaners.Application.Commands;
 public class DeleteCleanerCommandHandler(ILogger<DeleteCleanerCommandHandler> logger,
-    ICleanersRepository cleanersRepository) : IRequestHandler<DeleteCleanerCommand>
+    ICleanersRepository cleanersRepository,
+    ICleanerAuthorizationService cleanerAuthorizationService) : IRequestHandler<DeleteCleanerCommand>
 {
     public async Task Handle(DeleteCleanerCommand request, CancellationToken cancellationToken)
     {
@@ -18,6 +20,12 @@ public class DeleteCleanerCommandHandler(ILogger<DeleteCleanerCommandHandler> lo
         { 
             throw new NotFoundException(nameof(Cleaner), request.Id.ToString()); 
         }
+
+        if (!cleanerAuthorizationService.Authorize(cleaner, Domain.Constants.ResourceOperation.Delete)) 
+        {
+            throw new ForbidException();
+        }
+
         await cleanersRepository.Delete(cleaner);
     }
 }
